Optimización Custom error

+++ src/infrastructure/datasources/action.datasource.impl.ts
@@ -1,20 +1,20 @@
+import { ActionsModel } from "../../data/mongodb";
 import { ActionDatasource } from "../../domain/datasources";
 import { AddActionDto, DeleteActionDto } from "../../domain/dtos/roles-permissions";
 import { ActionEntity, DeleteActionEntity } from "../../domain/entities";
-import { ActionsModel } from "../../data/mongodb";
-import { ActionMapper } from "../mappers";
+import { CustomError } from "../../domain/errors/custom.error";
 
 export class ActionDatasourceImpl implements ActionDatasource {
 
      async addAction(addActionDto: AddActionDto): Promise<ActionEntity> {
-          const { actionName } = addActionDto;
+          const { actionName, hasError, errorMessage } = addActionDto;
           try {
                // 1. Crear el action
                const action = await ActionsModel.create({ actionName: actionName })
                await action.save();
 
                // 2. Mapear la respuesta a la entidad
-               return ActionMapper.actionEntityFromObject(action);
+               return new ActionEntity(action.id, actionName, hasError, errorMessage);
 
           } catch (error) {
                throw error;
@@ -22,21 +22,16 @@ export class ActionDatasourceImpl implements ActionDatasource {
      }
 
      async deleteAction(deleteActionDto: DeleteActionDto): Promise<DeleteActionEntity> {
-          const { actionName } = deleteActionDto;
+          const { actionName, hasError, errorMessage } = deleteActionDto;
           let message: string = "No hay action que coincida con la consulta. Se eliminaron 0 actions."
           try {
-               // 2. eliminar el action
-               const result = await ActionsModel.deleteOne({ actionName: actionName })
+               const exists = await ActionsModel.findOne({ actionName: actionName })
+               if (!exists) throw CustomError.badRequest("The role name does not exist or has been deleted.")
 
-               if (result.deletedCount === 1) {
-                    message = "Action deleted successfully.";
-               }
-               else {
-                    throw message
-               }
+               // 2. eliminar el role
+               await ActionsModel.deleteOne({ actionName: actionName })
 
-               // 3. Mapear la respuesta a la entidad
-               return ActionMapper.deleteActionEntityFromObject({ actionName, message });
+               return new DeleteActionEntity(actionName, hasError, errorMessage);
 
           } catch (error) {
                throw error;
